+++ pons/_client.py
@@ -352,6 +352,16 @@ class ClientSession:
         )
         return rpc_decode_quantity(result)
 
+    @rpc_call("eth_getCode")
+    async def eth_get_code(
+        self, address: Address, block: Union[int, Block] = Block.LATEST
+    ) -> bytes:
+        """Calls the ``eth_getCode`` RPC method."""
+        result = await self._provider_session.rpc(
+            "eth_getCode", address.rpc_encode(), rpc_encode_block(block)
+        )
+        return rpc_decode_data(result)
+
     async def wait_for_transaction_receipt(
         self, tx_hash: TxHash, poll_latency: float = 1.0
     ) -> TxReceipt: