+++ docs/changelog.rst
@@ -28,6 +28,7 @@ Added
 - ``LocalProvider.add_account()`` method. (PR_62_)
 - An optional ``sender_address`` parameter of ``ClientSession.eth_call()``. (PR_62_)
 - Expose ``Provider`` at the top level. (PR_63_)
+- ``eth_getCode`` support (as ``ClientSession.eth_get_code()``). (PR_64_)
 
 
 Fixed
@@ -46,6 +47,7 @@ Fixed
 .. _PR_61: https://github.com/fjarri/pons/pull/61
 .. _PR_62: https://github.com/fjarri/pons/pull/62
 .. _PR_63: https://github.com/fjarri/pons/pull/63
+.. _PR_64: https://github.com/fjarri/pons/pull/64
 
 
 0.7.0 (09-07-2023)